+++ spec/02-integration/05-proxy/23-context_spec.lua
@@ -1,10 +1,51 @@
 local helpers = require "spec.helpers"
 local null = ngx.null
+local atc_compat = require "kong.router.compat"
 
 
+local function reload_router(flavor)
+  _G.kong = {
+    configuration = {
+      router_flavor = flavor,
+    },
+  }
+
+  helpers.setenv("KONG_ROUTER_FLAVOR", flavor)
+
+  package.loaded["spec.helpers"] = nil
+  package.loaded["kong.global"] = nil
+  package.loaded["kong.cache"] = nil
+  package.loaded["kong.db"] = nil
+  package.loaded["kong.db.schema.entities.routes"] = nil
+  package.loaded["kong.db.schema.entities.routes_subschemas"] = nil
+
+  helpers = require "spec.helpers"
+
+  helpers.unsetenv("KONG_ROUTER_FLAVOR")
+end
+
+
+local function gen_route(flavor, r)
+  if flavor ~= "expressions" then
+    return r
+  end
+
+  r.expression = atc_compat.get_expression(r)
+  r.priority = tonumber(atc_compat._get_priority(r))
+
+  r.paths = nil
+  r.destinations = nil
+
+  return r
+end
+
+
+for _, flavor in ipairs({ "traditional", "traditional_compatible", "expressions" }) do
 for _, strategy in helpers.each_strategy() do
-  describe("Context Tests [#" .. strategy .. "]", function()
+  describe("Context Tests [#" .. strategy .. ", flavor = " .. flavor .. "]", function()
     describe("[http]", function()
+      reload_router(flavor)
+
       local proxy_client
       lazy_setup(function()
         local bp = helpers.get_db_utils(strategy, {
@@ -16,9 +57,9 @@ for _, strategy in helpers.each_strategy() do
           "ctx-tests-response",
         })
 
-        local unbuff_route = bp.routes:insert {
+        local unbuff_route = bp.routes:insert(gen_route(flavor, {
           paths   = { "/" },
-        }
+        }))
 
         bp.plugins:insert {
           name = "ctx-tests",
@@ -33,9 +74,9 @@ for _, strategy in helpers.each_strategy() do
           }
         }
 
-        local buffered_route = bp.routes:insert {
+        local buffered_route = bp.routes:insert(gen_route(flavor, {
           paths   = { "/buffered" },
-        }
+        }))
 
         bp.plugins:insert {
           name = "ctx-tests",
@@ -50,9 +91,9 @@ for _, strategy in helpers.each_strategy() do
           }
         }
 
-        local response_route = bp.routes:insert {
+        local response_route = bp.routes:insert(gen_route(flavor, {
           paths = { "/response" },
-        }
+        }))
 
         bp.plugins:insert {
           name = "ctx-tests-response",
@@ -68,6 +109,7 @@ for _, strategy in helpers.each_strategy() do
         }
 
         assert(helpers.start_kong({
+          router_flavor = flavor,
           database      = strategy,
           plugins       = "bundled,ctx-tests,ctx-tests-response",
           nginx_conf    = "spec/fixtures/custom_nginx.template",
@@ -128,6 +170,8 @@ for _, strategy in helpers.each_strategy() do
 
     if strategy ~= "off" then
       describe("[stream]", function()
+        reload_router(flavor)
+
         local MESSAGE = "echo, ping, pong. echo, ping, pong. echo, ping, pong.\n"
         local tcp_client
         lazy_setup(function()
@@ -145,7 +189,7 @@ for _, strategy in helpers.each_strategy() do
             protocol = "tcp",
           })
 
-          assert(bp.routes:insert {
+          assert(bp.routes:insert(gen_route(flavor, {
             destinations = {
               { port = 19000 },
             },
@@ -153,7 +197,7 @@ for _, strategy in helpers.each_strategy() do
               "tcp",
             },
             service = service,
-          })
+          })))
 
           bp.plugins:insert {
             name = "ctx-tests",
@@ -166,6 +210,7 @@ for _, strategy in helpers.each_strategy() do
           }
 
           assert(helpers.start_kong({
+            router_flavor = flavor,
             database      = strategy,
             stream_listen = helpers.get_proxy_ip(false) .. ":19000",
             plugins       = "bundled,ctx-tests",
@@ -197,3 +242,4 @@ for _, strategy in helpers.each_strategy() do
     end
   end)
 end
+end   -- for flavor