+++ spec/02-integration/05-proxy/28-stream_plugins_triggering_spec.lua
@@ -1,6 +1,7 @@
 local helpers = require "spec.helpers"
 local pl_file = require "pl.file"
 local cjson = require "cjson"
+local atc_compat = require "kong.router.compat"
 
 
 local TEST_CONF = helpers.test_conf
@@ -69,8 +70,48 @@ local function assert_phases(phrases)
   end
 end
 
+
+local function reload_router(flavor)
+  _G.kong = {
+    configuration = {
+      router_flavor = flavor,
+    },
+  }
+
+  helpers.setenv("KONG_ROUTER_FLAVOR", flavor)
+
+  package.loaded["spec.helpers"] = nil
+  package.loaded["kong.global"] = nil
+  package.loaded["kong.cache"] = nil
+  package.loaded["kong.db"] = nil
+  package.loaded["kong.db.schema.entities.routes"] = nil
+  package.loaded["kong.db.schema.entities.routes_subschemas"] = nil
+
+  helpers = require "spec.helpers"
+
+  helpers.unsetenv("KONG_ROUTER_FLAVOR")
+end
+
+
+local function gen_route(flavor, r)
+  if flavor ~= "expressions" then
+    return r
+  end
+
+  r.expression = atc_compat.get_expression(r)
+  r.priority = tonumber(atc_compat._get_priority(r))
+
+  r.destinations = nil
+
+  return r
+end
+
+
+for _, flavor in ipairs({ "traditional", "traditional_compatible", "expressions" }) do
 for _, strategy in helpers.each_strategy() do
-  describe("#stream Proxying [#" .. strategy .. "]", function()
+  describe("#stream Proxying [#" .. strategy .. ", flavor = " .. flavor .. "]", function()
+    reload_router(flavor)
+
     local bp
 
     before_each(function()
@@ -89,7 +130,7 @@ for _, strategy in helpers.each_strategy() do
         protocol = "tcp"
       })
 
-      bp.routes:insert {
+      bp.routes:insert(gen_route(flavor, {
         destinations = {
           {
             port = 19000,
@@ -99,7 +140,7 @@ for _, strategy in helpers.each_strategy() do
           "tcp",
         },
         service = tcp_srv,
-      }
+      }))
 
       local tls_srv = bp.services:insert({
         name = "tls",
@@ -108,7 +149,7 @@ for _, strategy in helpers.each_strategy() do
         protocol = "tls"
       })
 
-      bp.routes:insert {
+      bp.routes:insert(gen_route(flavor, {
         destinations = {
           {
             port = 19443,
@@ -118,13 +159,14 @@ for _, strategy in helpers.each_strategy() do
           "tls",
         },
         service = tls_srv,
-      }
+      }))
 
       bp.plugins:insert {
         name = "logger",
       }
 
       assert(helpers.start_kong({
+        router_flavor = flavor,
         database   = strategy,
         nginx_conf = "spec/fixtures/custom_nginx.template",
         plugins = "logger",
@@ -163,7 +205,9 @@ for _, strategy in helpers.each_strategy() do
     end)
   end)
 
-  describe("#stream Proxying [#" .. strategy .. "]", function()
+  describe("#stream Proxying [#" .. strategy .. ", flavor = " .. flavor .. "]", function()
+    reload_router(flavor)
+
     local bp
 
     before_each(function()
@@ -183,7 +227,7 @@ for _, strategy in helpers.each_strategy() do
         protocol = "tcp"
       })
 
-      bp.routes:insert {
+      bp.routes:insert(gen_route(flavor, {
         destinations = {
           {
             port = 19000,
@@ -193,7 +237,7 @@ for _, strategy in helpers.each_strategy() do
           "tcp",
         },
         service = tcp_srv,
-      }
+      }))
 
       local tls_srv = bp.services:insert({
         name = "tls",
@@ -202,7 +246,7 @@ for _, strategy in helpers.each_strategy() do
         protocol = "tls"
       })
 
-      bp.routes:insert {
+      bp.routes:insert(gen_route(flavor, {
         destinations = {
           {
             port = 19443,
@@ -212,7 +256,7 @@ for _, strategy in helpers.each_strategy() do
           "tls",
         },
         service = tls_srv,
-      }
+      }))
 
       bp.plugins:insert {
         name = "logger",
@@ -227,6 +271,7 @@ for _, strategy in helpers.each_strategy() do
       }
 
       assert(helpers.start_kong({
+        router_flavor = flavor,
         database   = strategy,
         nginx_conf = "spec/fixtures/custom_nginx.template",
         plugins = "logger,short-circuit",
@@ -277,3 +322,4 @@ for _, strategy in helpers.each_strategy() do
     end)
   end)
 end
+end -- for flavor